Fixed the code that reorders columns
The code to reorder columns was using cut paste except it cutting everything regardless if it needed to be moved. Cut is still not a good method because it tends to skip columns. The best solution ended up being to make sure the columns were ordered correctly in the first place. I also made a small edit to an autofilter function which may need further enhancing in the future. Note: with this data, I got an "FYI" which Camille placed for deleted rows. This is most likely a glitch. The output looked good, but since the columns aren't ordered exactly the same as they used to, we may run into issues down the line. Just a reminder not to rely on "hard code"