Important! Fixed UICPM to accept all years of truck data and fixed lat long problem
Warning! See lines 1033-1036 of DP_10_RoadInt.bas. This modifies the percent trucks with a somewhat complicated equation. I adapted it to work for all years of truck data but I'm not sure what it does. We need to make sure it wasn't designed specifically for using a single year of truck data because if it is, we should change it. The equation is something like

Ti = (Ti * AADTi + New_Ti * New_AADTi) / (New_AADTi + AADTi)

Where T is percent trucks. I don't really know what changes with the new T and AADT but it seems suspicious.