fixed fix_endpoints code, and the compress code at the end.
It turns out they already made code that compresses segments like the code we made lol. It just needed to be revised to our column numbers. That being said, there was still compression done after the tables were combined which tells me something might be up with the shell code.

Also, there are some segments that we probably won't be able to fix as explained by the warning messages, but at least the number of problem segments has been greatly reduced. The problem may also be with the route file we are reading in.