WARNING: This is a very questionable method to remove non-existent route/years.
This is especially problematic for intersections because the input intersection file seems to be missing route approach information which could lead to the missing aadt data as well. However, this works for the most part because we modified the AADT file to estimate missing data as much as we felt was reasonable. Remember that we do not estimate missing AADT if all previous years were also missing data. This makes the assumption that the segment did not exist before that year.