Improved code to add medians
Instead of using median midpoints I check if the median begins before the segment ends and ends after the segment begins. This ensures we get all medians even partially contained in the segment. Then I determine internal length of these medians and accumulate any medians of the same type. FInally, we return the median of the max length using which.max()