code to add several median columns as described in our meeting today
This code is flexible enough to turn into a general function easily if we wanted to use it on other datasets.

Note: there is a parameter called "min_portion" which tells the function to exclude medians which take up less than that portion of the segment. I found that this needs to be very small or it will cause additional segments to return nothing for the medians. This implies that those segments are "bookended" by overlapping medians, but essentially contain no median data otherwise. Even if we try to maintain these "bookends" there are still 40 segments without medians.