Finish up intersection fixes
I finished making the changes described in the previous commit. However, we still need to determine what to do about missing negative route data. My thoughts are that we merge negative routes with positive routes in the input files and if there are any negative AADTs that are only in one direction that we multiply it by two so that it works with the existing function to add AADT.