Modified functional area code to meet UDOT standards
Functional area no longer relies on speed limit. Instead, it pulls from the intersection type and stop control only. I made the assumption that the values given on the website are to be measured from the center of the intersection. The UDOT functional area standards are explained on this website.
https://sites.google.com/utah.gov/udot-safety-standards/safety-evaluations/historic-and-predictive-safety-analysis?authuser=0