Increasing lat long precision for intersections
I went through a lot of effort to join higher precision lat long data to the interesection dataframe, but it was creating duplicates. I finally realized the intersection file contained a "lat_text" and "long_text" which had 6 decimal points instead of 3. This makes a huge difference, so I didn't really need to do all that :(